chore: move internal tools to root scripts
