+++ package.json
@@ -6,7 +6,6 @@
     "packages/*",
     "services/*",
     "infra/*",
-    "tools",
     "docs"
   ],
   "scripts": {
@@ -24,9 +23,8 @@
     "consistency-check-licenses": "node .scripts/licenses.mjs consistency-check",
     "build:all": "nx run-many --targets=build",
     "build:shared": "tsdown --config shared/tsdown.config.ts",
-    "attw:all": "yarn tools attw",
+    "attw:all": "tsx .scripts/attw.ts",
     "clear": "rimraf --glob \"**/*/dist\" .nx",
-    "tools": "./bin/tools",
     "publish:dev": "sh .scripts/publish-dev.sh"
   },
   "resolutions": {
@@ -52,6 +50,7 @@
     "@types/react": "catalog:react-native",
     "@yarnpkg/types": "^4.0.0",
     "chalk": "^5.3.0",
+    "es-toolkit": "^1.39.10",
     "eslint": "^9.10.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
@@ -65,6 +64,7 @@
     "hygen": "6.2.11",
     "knip": "^1",
     "nx": "^21.0.3",
+    "picocolors": "^1.1.1",
     "prettier": "^3.3.3",
     "rimraf": "^6.0.1",
     "ts-node": "^10.9.2",
@@ -72,6 +72,7 @@
     "tsx": "^4.19.3",
     "typescript": "catalog:tools",
     "typescript-eslint": "^8.31.0",
-    "yargs": "^17.7.2"
+    "yargs": "^17.7.2",
+    "zx": "^8.5.4"
   }
 }
