--- tools/README.md
@@ -1,3 +1,0 @@
-# tools
-
-A set of commands for Granite.
