--- tools/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@granite-internal/tools",
-  "private": true,
-  "type": "module",
-  "bin": {
-    "granite-tools": "./dist/index.js"
-  },
-  "scripts": {
-    "prepack": "yarn build",
-    "dev": "tsx src/index.ts --",
-    "typecheck": "tsc --noEmit",
-    "build": "tsup"
-  },
-  "main": "./dist/index.js",
-  "devDependencies": {
-    "@commander-js/extra-typings": "^13.1.0",
-    "@nx/devkit": "^21.0.3",
-    "@types/node": "catalog:tools",
-    "@types/yargs": "^17.0.33",
-    "@types/yauzl": "^2.10.3",
-    "commander": "^13.1.0",
-    "es-toolkit": "^1.39.10",
-    "nx": "^21.0.3",
-    "picocolors": "^1.1.1",
-    "tsup": "^8.5.0",
-    "tsx": "^4.19.4",
-    "typescript": "catalog:tools",
-    "zx": "^8.5.4"
-  }
-}
